+++ test/choice.js
@@ -49,7 +49,9 @@ describe("test/choice.js", function () {
 	});
 
 	describe("With an non-random distribution, each path is called equally", function () {
-		state.random = randRobin;
+		var oldRandom = state.random;
+
+		state.setRandom(randRobin);
 
 		var instance2 = new state.StateMachineInstance("instance2");
 		instance2.path1 = 0;
@@ -69,5 +71,7 @@ describe("test/choice.js", function () {
 			assert.equal(33, instance2.path2);
 			assert.equal(33, instance2.path3);
 		});
+
+		state.setRandom(oldRandom);
 	});
 });